+++ src/extension/app/components/plugin/plugin-action-bar.js
@@ -132,7 +132,7 @@ export class PluginActionBar extends ConnectedElement {
     systemPlugins.push(ICONS.ADOBE_LOGO);
 
     const actionGroup = html`<sp-action-group>${systemPlugins}</sp-action-group>`;
-    const divider = loggedIn || siteStore.authorized ? html`<sp-divider size="s" vertical></sp-divider>` : '';
+    const divider = html`<sp-divider size="s" vertical></sp-divider>`;
 
     return [divider, actionGroup];
   }