+++ src/extension/ui.js
@@ -115,7 +115,7 @@ export async function updateContextMenu({
     // clear context menu
     await chrome.contextMenus.removeAll();
 
-    if (isValidProject(config)) {
+    if (isValidProject(config) && !url.startsWith(GH_URL)) {
       const { owner, repo } = config;
       const project = await getProject(`${owner}/${repo}`);
       // add/remove project config
@@ -139,15 +139,6 @@ export async function updateContextMenu({
           contexts: [
             'action',
           ],
-        });
-        // open preview url
-        await chrome.contextMenus.create({
-          id: 'openPreview',
-          title: chrome.i18n.getMessage('open_preview'),
-          contexts: [
-            'action',
-          ],
-          visible: url.startsWith(GH_URL),
         });
       }
       // open view doc source
@@ -177,6 +168,27 @@ export async function updateUI(context = {}) {
 // add listener for clicks on context menu item
 if (chrome.contextMenus) {
   chrome.contextMenus.onClicked.addListener(async ({ menuItemId }, tab) => {
+    if (!tab.url) return;
     internalActions[menuItemId](tab);
+    await chrome.scripting.executeScript({
+      target: { tabId: tab.id },
+      func: async (menuItemIdVal) => {
+        try {
+          const mod = await import(chrome.runtime.getURL('rum.js'));
+          const { default: sampleRUM } = mod;
+
+          // Ensure window.hlx and window.hlx.sidekick exists
+          window.hlx = window.hlx || {};
+          window.hlx.sidekick = window.hlx.sidekick || { location: window.location };
+
+          const checkpoint = `sidekick:context-menu:${menuItemIdVal}`;
+          sampleRUM(checkpoint, { source: window.location.href });
+        } catch (e) {
+          // eslint-disable-next-line no-console
+          console.log('Unable to collect RUM data', e);
+        }
+      },
+      args: [menuItemId],
+    });
   });
 }