+++ test/app/store/site.test.js
@@ -104,7 +104,7 @@ describe('Test Site Store', () => {
       expect(appStore.siteStore.devUrl.origin).to.equal('http://localhost:3000');
       expect(appStore.siteStore.views.length).to.equal(1);
       expect(appStore.siteStore.views[0].path).to.equal('**.json');
-      expect(appStore.siteStore.views[0].viewer).to.equal('/test/wtr/fixtures/views/json/json.html');
+      expect(appStore.siteStore.views[0].viewer).to.equal('/test/fixtures/views/json/json.html');
       expect(appStore.siteStore.plugins.length).to.equal(0);
     });
 
@@ -157,7 +157,7 @@ describe('Test Site Store', () => {
       expect(appStore.siteStore.views[0].viewer).to.equal('/tools/sidekick/example/index.html');
 
       expect(appStore.siteStore.views[1].path).to.equal('**.json');
-      expect(appStore.siteStore.views[1].viewer).to.equal('/test/wtr/fixtures/views/json/json.html');
+      expect(appStore.siteStore.views[1].viewer).to.equal('/test/fixtures/views/json/json.html');
       // @ts-ignore
       expect(appStore.siteStore.views[1].title()).to.equal('Data rendition');
     });