+++ test/fixtures/helix-admin.js
@@ -128,6 +128,16 @@ export const defaultConfigJSONWithUnpinnedPlugin = {
   ],
 };
 
+export const defaultEnvJSON = {
+  version: 1,
+  prod: {
+    host: 'foo.bar',
+    routes: [],
+  },
+  contentSourceUrl: 'https://adobe.sharepoint.com/sites/foo/Shared%20Documents/root',
+  contentSourceType: 'onedrive',
+};
+
 /**
  * status request stubs
  */