+++ test/ui.test.js
@@ -11,40 +11,42 @@
  */
 /* eslint-disable no-unused-expressions */
 
-import { expect } from '@open-wc/testing';
+import { aTimeout, expect } from '@open-wc/testing';
 import { setUserAgent } from '@web/test-runner-commands';
 import sinon from 'sinon';
 
 import chromeMock from './mocks/chrome.js';
 import { addProject, getProject, updateProject } from '../src/extension/project.js';
-import { internalActions } from '../src/extension/actions.js';
 import { setDisplay } from '../src/extension/display.js';
+import { internalActions } from '../src/extension/actions.js';
 import { error } from './test-utils.js';
+import { mockFetchEnvJSONSuccess } from './mocks/helix-admin.js';
 
 // @ts-ignore
 window.chrome = chromeMock;
 
-// prep listener test before importing ui.js
-const openPreviewSpy = sinon.spy(internalActions, 'openPreview');
-sinon.replace(chrome.contextMenus.onClicked, 'addListener', async (func) => {
-  // @ts-ignore
-  func({
-    menuItemId: 'openPreview',
-  }, {
-    id: 1,
-    url: 'https://main--blog--adobe.hlx.page/',
-  });
-});
-
-const { updateContextMenu, updateIcon } = await import('../src/extension/ui.js');
-
 const sandbox = sinon.createSandbox();
 const config = {
-  owner: 'foo',
-  repo: 'bar',
+  owner: 'adobe',
+  repo: 'aem-boilerplate',
   ref: 'main',
   mountpoints: [],
 };
+const url = 'https://main--blog--adobe.hlx.page/';
+const tab = {
+  id: 1,
+  url,
+};
+let clickListener;
+
+// prep listener test before importing ui.js
+const logSpy = sinon.spy(console, 'log');
+sinon.replace(chrome.contextMenus.onClicked, 'addListener', async (func) => {
+  // store click listener for later use
+  clickListener = func;
+});
+
+const { updateContextMenu, updateIcon } = await import('../src/extension/ui.js');
 
 describe('Test UI: updateContextMenu', () => {
   let createSpy;
@@ -70,12 +72,19 @@ describe('Test UI: updateContextMenu', () => {
   });
 
   it('click listener invokes action', async () => {
-    expect(openPreviewSpy.called).to.be.true;
+    const openViewDocSourceStub = sinon.stub(internalActions, 'openViewDocSource');
+    clickListener({
+      menuItemId: 'openViewDocSource',
+    }, {
+      id: 1,
+      url: 'https://main--blog--adobe.hlx.page/',
+    });
+    expect(openViewDocSourceStub.called).to.be.true;
   });
 
   it('updateContextMenu: project not added yet', async () => {
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
       config,
     });
     expect(removeAllSpy.callCount).to.equal(1);
@@ -85,31 +94,31 @@ describe('Test UI: updateContextMenu', () => {
   });
 
   it('updateContextMenu: project added and enabled', async () => {
+    mockFetchEnvJSONSuccess();
     await addProject(config);
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
       config,
     });
     expect(removeAllSpy.callCount).to.equal(1);
-    expect(createSpy.callCount).to.equal(4);
+    expect(createSpy.callCount).to.equal(3);
     expect(createSpy.calledWithMatch({ id: 'enableDisableProject' })).to.be.true;
-    expect(createSpy.calledWithMatch({ id: 'openPreview' })).to.be.true;
   });
 
   it('updateContextMenu: project added and disabled', async () => {
     const project = await getProject(config);
     await updateProject({ ...project, disabled: true });
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
       config,
     });
     expect(removeAllSpy.callCount).to.equal(1);
-    expect(createSpy.callCount).to.equal(4);
+    expect(createSpy.callCount).to.equal(3);
   });
 
   it('updateContextMenu: no matching config', async () => {
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
     });
     expect(removeAllSpy.callCount).to.equal(1);
     expect(createSpy.callCount).to.equal(0);
@@ -132,25 +141,33 @@ describe('Test UI: updateContextMenu', () => {
         { tab: { id: 1 } },
       ));
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
       config,
     });
     expect(removeAllSpy.callCount).to.equal(1);
-    expect(createSpy.callCount).to.equal(3);
+    expect(createSpy.callCount).to.equal(2);
     expect(createSpy.calledWithMatch({ id: 'openViewDocSource' })).to.be.false;
   });
 
   it('updateContextMenu: guessAEMSite fails', async () => {
-    // addListenerStub.restore();
     sandbox.stub(chrome.scripting, 'executeScript')
       .throws(error);
     await updateContextMenu({
-      url: 'https://main--bar--foo.hlx.page/',
+      url,
       config,
     });
     expect(removeAllSpy.callCount).to.equal(1);
-    expect(createSpy.callCount).to.equal(3);
+    expect(createSpy.callCount).to.equal(2);
     expect(createSpy.calledWithMatch({ id: 'openViewDocSource' })).to.be.false;
+  });
+
+  it('updateContextMenu: ignore github url', async () => {
+    await updateContextMenu({
+      url: 'https://github.com/foo/bar/',
+      config,
+    });
+    expect(removeAllSpy.callCount).to.equal(1);
+    expect(createSpy.callCount).to.equal(0);
   });
 });
 
@@ -214,5 +231,40 @@ describe('Test UI: updateIcon', () => {
         512: 'icons/hidden/icon-512x512.png',
       },
     })).to.be.true;
+  });
+});
+
+describe('Test UI: RUM collection when clicked', () => {
+  before(async () => {
+    await setUserAgent('HeadlessChrome');
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it('collects RUM', async () => {
+    clickListener({
+      menuItemId: 'openViewDocSource',
+    }, tab);
+    await aTimeout(1000);
+    expect(logSpy.calledWithMatch('sampleRUM', 'sidekick:context-menu:openViewDocSource')).to.be.true;
+  });
+
+  it('handles error', async () => {
+    clickListener({
+      menuItemId: 'openViewDocSource', // provoke error in sampleRUM
+    }, tab);
+    await aTimeout(500);
+    expect(logSpy.calledWithMatch('Unable to collect RUM data', error)).to.be.true;
+  });
+
+  it('does nothing without tab url', async () => {
+    const logs = logSpy.callCount;
+    clickListener({
+      menuItemId: 'openViewDocSource',
+    }, {});
+    await aTimeout(500);
+    expect(logSpy.callCount).to.equal(logs);
   });
 });